changed back to using full rgb textures as using emulated vga color palette just unnecessarily makes things more complicated, but going to keep the code in commented out just in case

+++ src/display.h
@@ -9,7 +9,7 @@ extern int width, height, size;
 extern uint32_t *pixels;
 } // namespace display
 
-void AddPixelToBuffer(int x, int y, uint8_t color);
-void AddPixelToBufferUnsafe(int x, int y, uint8_t color);
+void AddPixelToBuffer(int x, int y, uint32_t color);
+void AddPixelToBufferUnsafe(int x, int y, uint32_t color);
 
 #endif
